Eliminación realizada con exito. Curso Finalizado

+++ src/app/home/home.component.ts
@@ -10,6 +10,10 @@ import { Movie } from '../interfaces/movie';
 export class HomeComponent implements OnInit {
   movies: Movie[] = [];
   constructor(private movieService: MoviesService) {
+    this.getMovies();
+  }
+  
+  getMovies(){
     this.movieService.get().subscribe( (data: any) => {// En lugar de 'any' era 'Movies[]'
       this.movies = data;
     }, (error) => {
@@ -17,16 +21,18 @@ export class HomeComponent implements OnInit {
       alert('Ocurrio un error');
     });
   }
-
   ngOnInit(): void {
   }
 
   delete(id:any){
-    this.movieService.delete(id).subscribe( (data) => {
-      alert("Eliminado con éxito");
-    }, (error) => {
-      console.log(error);
-      alert('Ocurrio un error');
-    });
+    if( confirm('Seguro que deseas eliminar esta pelicula?')){
+      this.movieService.delete(id).subscribe( (data) => {
+        alert("Eliminado con éxito");
+        this.getMovies();
+      }, (error) => {
+        console.log(error);
+        alert('Ocurrio un error');
+      });
+    }
   }
 }
